requirements.txt
gradio
pillow
numpy
opencv-python
torch
transformers>=4.41
accelerate
sentencepiece
torchvision
